Add test data for connections of example plant
The problem of dependency fixtures being nonexistent was encountered and
was assumed to stem from the fixure files being executed in alphabetical
order.

To avoid this problem there were two options.
Here the option selected (due to small manageable data size and sunk
cost of having created the yaml anyhow) was that the correct ids were
hardcoded to match the ids already manually allocated in the segments
fixture yaml.